Move rendering to own module
