+++ src/modules/gui.cpp
@@ -1,10 +1,10 @@
 #include "gui.h"
 #include "backends/imgui_impl_sdl2.h"
 #include "backends/imgui_impl_sdlrenderer2.h"
-#include "flecs/addons/cpp/mixins/pipeline/decl.hpp"
-#include "graphics.h"
 #include "imgui.h"
+#include "modules/render.h"
 #include "spdlog/spdlog.h"
+#include <flecs.h>
 
 void systemInitialiseGui(flecs::iter &iter);
 void systemGuiNewFrame(flecs::iter &);
@@ -12,6 +12,8 @@ void systemGuiEndFrame(flecs::iter &);
 void systemRenderGUI(const Renderer &);
 
 GuiModule::GuiModule(flecs::world &world) {
+  world.import <RenderModule>();
+
   flecs::entity PreFramePhase = world.lookup("Phase.PreFrame");
   flecs::entity PreRenderPhase = world.lookup("Phase.PreRender");
   flecs::entity RenderPhase = world.lookup("Phase.Render");
